Refactor of list code & pie chart and only displaying the user entries for the chart and list
The code had to be refactored again so that the transaction list automatically updates without having to relaunch the application. It also had to be refactor due to an error where it was located in a stateless widget and could no longer be edited. And the pie chart and list finally now only display their entries in the database and no longer the entire database. And a few other fixes were made to the code